this is a bit of an optimization hack.
the problem is that reference counting proxy objects are pool allocated,
but the size of the pool grows linearly. If you need millions of reference
counted things all of the sudden, you can spend minutes and minutes
inside of realloc as the pool slowly grows to meet demand. this solves
my problem by making the slope of the line a whole lot steeper, but
it probably is a bit agressive in memory usage and it still will be
really slow if you happen to need yet more reference counted things
than I do. The clear solution here is to not use a linear growth
for the pool size, but that's work I'm not sure I understand this
code well enough to do in the alloted time, so I'll leave it to someone
else with a bit better understanding of what's going on here. for now,
I'm forced to commit the hack because this is a submodule in a larger
project and I need to distribute the fix.